+++ CHANGES.rst
@@ -115,6 +115,9 @@ New:
 - Add ``defaultSortOn`` option in ``livesearch`` pattern.
   [Gagaro]
 
+- Add an option to set the path operator in QueryHelp and tinymce pattern.
+  [Gagaro]
+
 Fixes:
 
 - Change ``bool`` function in mockup-utils to allow for truthy values and match on falsy values.